Refactoring thunks in different files; drag and drop fixed
