+++ src/components/FileExplorer/FileExplorerItem/FileListItem/FileListItem.scss
@@ -3,7 +3,7 @@
     .file-list-item {
         background-color: white;
     
-        @apply py-3.5 border-b border-l-neutral-30 text-sm w-full flex;
+        @apply py-3.5 border-b border-l-neutral-30 text-base w-full flex;
 
         &:hover, &.selected {
             @apply bg-blue-20;
@@ -19,7 +19,7 @@
     }
 
     .file-list-item-name-span {
-        @apply max-w-full whitespace-nowrap overflow-hidden overflow-ellipsis text-neutral-900 text-sm px-1 border border-dashed border-opacity-0 border-blue-40 hover:border-opacity-100;
+        @apply max-w-full whitespace-nowrap overflow-hidden overflow-ellipsis text-neutral-900 px-1 border border-dashed border-opacity-0 border-blue-40 hover:border-opacity-100;
     }
 
     .file-list-item-actions-button {
