+++ src/components/FileExplorer/FilesList/FilesList.tsx
@@ -65,7 +65,7 @@ class FilesList extends React.Component<FilesListProps, FilesListState> {
     return (
       <div className="flex-grow bg-white">
         <div className="w-full">
-          <div className="files-list flex border-b border-l-neutral-30 bg-white text-neutral-500 py-2 text-sm">
+          <div className="files-list font-semibold flex border-b border-l-neutral-30 bg-white text-neutral-500 py-2 text-sm">
             <div className="w-0.5/12 px-3 flex items-center justify-center box-content">
               <input disabled={!this.hasItems} readOnly checked={this.isAllSelected} onClick={this.onSelectAllButtonClicked} type="checkbox" className="pointer-events-auto" />
             </div>
