+++ src/index.scss
@@ -165,11 +165,15 @@ code {
     min-width: 150px;
     max-width: 250px;
 
-    @apply no-underline py-1 px-2 hover:bg-l-neutral-20 active:bg-l-neutral-30 rounded-1px text-neutral-900 text-sm flex items-center;
+    @apply no-underline py-1 px-2 hover:bg-l-neutral-20 active:bg-l-neutral-30 rounded-1px text-neutral-900 text-base flex items-center;
 
     img {
       height: 14px;
     }
+  }
+
+  .dropdown-item+.dropdown-item {
+    @apply mt-2;
   }
 
   /* INPUTS */
