+++ src/services/i18n.service/locales/en.json
@@ -18,6 +18,7 @@
     "changingName": "Error during name change",
     "fileAlreadyExists": "File with the same name already exists",
     "folderAlreadyExists": "Folder with the same name already exists",
+    "creatingFolder": "Error during folder creation",
     "sendReferralInvitation": "Could not send invitation email",
     "sendClaimEmail": "Could not send claim email",
     "anyReferralCredit": "You don\t have any credit on your account",
