+++ src/store/slices/storage/thunks/createFolderTreeStructureThunk.ts
@@ -1,0 +1,108 @@
+import { ActionReducerMapBuilder, createAsyncThunk } from '@reduxjs/toolkit';
+
+import { StorageState } from '..';
+import { RootState } from '../../..';
+import notify, { ToastType } from '../../../../components/Notifications';
+import { FileActionTypes, FileStatusTypes } from '../../../../models/enums';
+import { NotificationData } from '../../../../models/interfaces';
+import folderService from '../../../../services/folder.service';
+import i18n from '../../../../services/i18n.service';
+import { tasksActions } from '../../tasks';
+import { selectorIsTeam } from '../../team';
+import { uploadItemsThunk } from './uploadItemsThunk';
+
+interface IRoot extends DirectoryEntry {
+  folderId?: number;
+  childrenFiles: File[],
+  childrenFolders: IRoot[],
+  fullPathEdited: string
+}
+
+interface CreateFolderTreeStructurePayload {
+  root: IRoot,
+  currentFolderId: number,
+  options?: {
+    withNotification?: boolean;
+    onSuccess?: () => void;
+  }
+}
+
+export const createFolderTreeStructureThunk = createAsyncThunk<void, CreateFolderTreeStructurePayload, { state: RootState }>(
+  'storage/createFolderStructure',
+  async ({ root, currentFolderId, options }, { getState, dispatch, requestId }) => {
+    const isTeam: boolean = selectorIsTeam(getState());
+    const levels = [root];
+    const notification: NotificationData = {
+      uuid: requestId,
+      action: FileActionTypes.UploadFolder,
+      status: FileStatusTypes.Pending,
+      name: root.name,
+      isFolder: true
+    };
+
+    options = Object.assign({ withNotification: true }, options || {});
+
+    if (options?.withNotification) {
+      dispatch(tasksActions.addNotification(notification));
+    }
+
+    try {
+      root.folderId = currentFolderId;
+
+      while (levels.length > 0) {
+        const level: IRoot = levels.shift() as IRoot;
+        const folderUploaded = await folderService.createFolder(isTeam, level.folderId, level.name);
+
+        await dispatch(uploadItemsThunk({
+          files: level.childrenFiles || [],
+          parentFolderId: folderUploaded.id,
+          folderPath: level.fullPathEdited,
+          options: { withNotifications: false }
+        }));
+
+        for (const child of level.childrenFolders) {
+          child.folderId = folderUploaded.id;
+        }
+
+        levels.push(...level.childrenFolders);
+      }
+
+      if (options?.withNotification) {
+        dispatch(tasksActions.updateNotification({
+          uuid: notification.uuid,
+          merge: {
+            status: FileStatusTypes.Success
+          }
+        }));
+      }
+
+      options.onSuccess?.();
+    } catch (error) {
+      options?.withNotification && dispatch(tasksActions.updateNotification({
+        uuid: notification.uuid,
+        merge: {
+          status: FileStatusTypes.Error
+        }
+      }));
+
+      throw error;
+    }
+  }
+);
+
+export const createFolderTreeStructureThunkExtraReducers = (builder: ActionReducerMapBuilder<StorageState>): void => {
+  builder
+    .addCase(createFolderTreeStructureThunk.pending, (state, action) => { })
+    .addCase(createFolderTreeStructureThunk.fulfilled, (state, action) => { })
+    .addCase(createFolderTreeStructureThunk.rejected, (state, action) => {
+      let errorMessage = i18n.get('error.uploadingFolder');
+
+      if (action.error.message?.includes('already exists')) {
+        errorMessage = i18n.get('error.folderAlreadyExists');
+      } else {
+        errorMessage = action.error.message || action.error + '';
+      }
+
+      notify(errorMessage, ToastType.Error);
+    });
+};
